backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyodbc==5.2.0
python-dotenv==1.0.1
